Reorganize codebase: Create bookextract library package
- Create bookextract/ package directory with proper __init__.py
- Move library modules into package:
  - book_capture.py -> bookextract/book_capture.py
  - book_intermediate.py -> bookextract/book_intermediate.py
  - ocr_processor.py -> bookextract/ocr_processor.py
- Extract and move reusable classes to library:
  - ImageProcessor from crop_gui.py -> bookextract/image_processor.py
  - EpubGenerator from render_epub.py -> bookextract/epub_generator.py
- Update all import statements to use package imports
- Remove shebang lines from library modules (keep in main programs)
- Add graceful handling of optional dependencies (ebooklib)
- Update README.md to reflect new package structure

Benefits:
- Clean separation between library code and GUI applications
- Reusable components for future development
- Better code organization and maintainability
- Proper Python package structure
