+++ bookextract/__init__.py
@@ -1,0 +1,37 @@
+from .book_capture import BookCapture
+from .book_intermediate import (
+    BookIntermediate, 
+    BookConverter, 
+    BookMetadata, 
+    Chapter, 
+    ContentSection,
+)
+from .intermediate_to_m4b import (
+    create_text_files_from_intermediate,
+    clean_text_for_tts,
+    create_metadata_file,
+    process_intermediate_file,
+    process_legacy_format,
+    process_intermediate_file_object,
+)
+from .ocr_processor import OCRProcessor
+from .image_processor import ImageProcessor
+from .epub_generator import EpubGenerator
+
+__all__ = [
+    'EpubGenerator',
+    'BookCapture',
+    'BookIntermediate',
+    'BookConverter', 
+    'BookMetadata',
+    'Chapter',
+    'ContentSection',
+    'OCRProcessor',
+    'ImageProcessor',
+    'create_text_files_from_intermediate',
+    'clean_text_for_tts',
+    'create_metadata_file',
+    'process_intermediate_file',
+    'process_legacy_format',
+    'process_intermediate_file_object',
+]
