+++ bookextract/ocr_processor.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """
 OCR Processing Module
